requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mysql-connector-python==8.0.13
numpy==1.15.4
pandas==0.23.4
protobuf==3.6.1
python-dateutil==2.7.5
pytz==2018.9
requests==2.8.1
scikit-learn==0.20.2
scipy==1.2.0
sentencepiece==0.1.7
six==1.12.0
Werkzeug==0.14.1
